Restructure Nodes so they are passed as templates.
This allows the original parts form BinaryTree to be shared across class, but increase performance.

It also removes the add need to have both public left and right variables, and getLeft and getRight.

An attempt was made at only functions, but that ended up showing a surprisingly high performance hit.

So this may struggle with using an array implementation.