Remove c++2a <=> support.
The goal of adding support was to make it so the class could inheriently call std::string.compare() in a way class independent way.

<=> Offered this, but because I want this to have a continous interface for c++17 and c++20, I would have needed to add a templated type to make this work.

But because the template would need to explicitly passed, it would just "start working" and so I'm abandoning it.

If c++20 adds the ability to type cast from int to comparable result, then this would have had a chance.